+++ src/main/java/com/dao/CartItemDao.java
@@ -6,7 +6,7 @@ import com.model.CartItem;
 public interface CartItemDao {
 
 	void addCartItem(CartItem cartItem);
-	void removeCartItem(Long CartItemId);
+	void removeCartItem(int CartItemId);
 	void removeAllCartItems(Cart cart);
 
 }
