+++ src/main/java/com/model/CartItem.java
@@ -20,7 +20,7 @@ public class CartItem implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Long cartItemId;
+	private int cartItemId;
 
 	private int quality;
 
@@ -35,11 +35,11 @@ public class CartItem implements Serializable {
 	@JsonIgnore
 	private Cart cart;
 
-	public Long getCartItemId() {
+	public int getCartItemId() {
 		return cartItemId;
 	}
 
-	public void setCartItemId(Long cartItemId) {
+	public void setCartItemId(int cartItemId) {
 		this.cartItemId = cartItemId;
 	}
