+++ src/main/java/com/service/CustomerOrderService.java
@@ -5,5 +5,5 @@ import com.model.CustomerOrder;
 public interface CustomerOrderService {
 
 	void addCustomerOrder(CustomerOrder customerOrder);
-	double getCustomerOrderGrandTotal(Long cartId);
+	double getCustomerOrderGrandTotal(int cartId);
 }
